# docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_alemancenter_fiber-api_internal_services.RedisInfoResponse:
    properties:
      info:
        type: string
    type: object
  github_com_alemancenter_fiber-api_internal_services.RedisKeysResponse:
    properties:
      count:
        type: integer
      keys:
        items:
          type: string
        type: array
    type: object
...
  github_com_alemancenter_fiber-api_internal_utils.APIResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.PaginationMeta'
      pagination:
        $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.PaginationMeta'
      success:
        type: boolean
      token:
        type: string
    type: object
  github_com_alemancenter_fiber-api_internal_utils.PaginationMeta:
    properties:
      current_page:
        type: integer
      from:
        type: integer
      last_page:
        type: integer
      per_page:
        type: integer
      to:
        type: integer
      total:
        type: integer
    type: object
...
  internal_handlers_redis.SetRedisKeyRequest:
    properties:
      key:
        type: string
      ttl:
        description: seconds
        type: integer
      value:
        type: string
    required:
    - key
    - value
    type: object
...
host: api.alemancenter.com
info:
  contact:
    email: support@alemancenter.com
    name: API Support
  description: Backend API for Alemancenter Educational Platform.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Alemancenter API
  version: 2.0.0
paths:
...
  /dashboard/redis:
    post:
      consumes:
      - application/json
      description: Set a key-value pair in Redis with an optional TTL
      parameters:
      - description: Redis key payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers_redis.SetRedisKeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Set Redis Key
      tags:
      - Redis
  /dashboard/redis/{key}:
    delete:
      description: Delete a specific Redis key
      parameters:
      - description: Redis Key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete Redis Key
      tags:
      - Redis
  /dashboard/redis/env:
    post:
      consumes:
      - application/json
      description: Validate and update Redis connection settings (Requires restart
        to apply fully)
      parameters:
      - description: Key-Value pairs of Redis config
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Update Redis Config
      tags:
      - Redis
  /dashboard/redis/expired/clean:
    delete:
      description: Clean up expired keys from Redis
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Clean Expired Keys
      tags:
      - Redis
  /dashboard/redis/info:
    get:
      description: Returns internal Redis server information and statistics
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_services.RedisInfoResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Redis Info
      tags:
      - Redis
  /dashboard/redis/keys:
    get:
      description: Returns a list of Redis keys matching a specific pattern
      parameters:
      - description: Key pattern to match (e.g. *)
        in: query
        name: pattern
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_services.RedisKeysResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: List Redis Keys
      tags:
      - Redis
  /dashboard/redis/test:
    get:
      description: Ping the Redis server to check connectivity and health
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alemancenter_fiber-api_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Test Redis Connection
      tags:
      - Redis
...
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
  FrontendKeyAuth:
    description: Frontend identifier key for public endpoints.
    in: header
    name: X-Frontend-Key
    type: apiKey
swagger: "2.0"